actions: move action definition to ai directory
